chore: move well-known links to folder
