+++ src/main.ts
@@ -4,10 +4,11 @@ import { secureHeaders } from "hono/secure-headers";
 //import { sentry } from "npm:@hono/sentry";
 import { PORT } from "./constants.ts";
 import { DB, runMigration } from "./db/db.ts";
-import { createLnurlApp, createWellKnownApp } from "./lnurlp.ts";
+import { createLnurlApp } from "./lnurlp.ts";
 import { LOG_LEVEL, logger, loggerMiddleware } from "./logger.ts";
 import { NWCPool } from "./nwc/nwcPool.ts";
 import { createUsersApp } from "./users.ts";
+import { createLnurlWellKnownApp, createNostrWellKnownApp } from "./well-known/index.ts";
 
 await runMigration();
 
@@ -26,7 +27,8 @@ hono.use(secureHeaders());
   hono.use("*", sentry({ dsn: SENTRY_DSN }));
 }*/
 
-hono.route("/.well-known", createWellKnownApp(db));
+hono.route("/.well-known/lnurlp", createLnurlWellKnownApp(db));
+hono.route("/.well-known/nostr.json", createNostrWellKnownApp(db));
 hono.route("/lnurlp", createLnurlApp(db));
 hono.route("/users", createUsersApp(db, nwcPool));
