+++ src/well-known/index.ts
@@ -1,0 +1,7 @@
+import { createLnurlWellKnownApp } from "./lnurlp.ts";
+import { createNostrWellKnownApp } from "./nostr.ts";
+
+export {
+  createLnurlWellKnownApp,
+  createNostrWellKnownApp
+};
